gesture swiping bug fix + integration change.
ViewPager -> ViewPager2

***MainActivity.java***
replaced ViewPager with ViewPager2.
used TabLayOutMediator to attach tableLayout with viewPager
created a custom adapter customAdapter

IMPORTANT:
notice the syntax in the constructor of TabLayoutMediator.
it accepts newly instantiated inner class with override to onConfigureTab method.
this is used to give names to the tabs by position, notice the order in customAdapter.java createFragment method.

NEW:
***customAdapter.java***
inherits FragmentStateAdapter, required for using ViewPager2 with Fragments.
NOTE:
createFragment is similar to getItem in PagerAdapter.
getItemCount returns an immutable value, in case of change in number of items will need to be changed manually.